// src/app/interfaces/currency-converted.ts
export interface CurrencyConverted {
  meta?: {};
  data: {
    EUR: {
      code: string;
      value: number;
    };
    USD: {
      code: string;
      value: number;
    };
    JPY: {
      code: string;
      value: number;
    };
    GBP: {
      code: string;
      value: number;
    };
    AUD: {
      code: string;
      value: number;
    };
    CAD: {
      code: string;
      value: number;
    };
    CHF: {
      code: string;
      value: number;
    };
    CNY: {
      code: string;
      value: number;
    };
    HKD: {
      code: string;
      value: number;
    };
  };
}